move inner static classes to toplevel package protected
