+++ gax/src/main/java/com/google/api/gax/rpc/ServerStream.java
@@ -31,11 +31,7 @@ package com.google.api.gax.rpc;
 
 import com.google.api.core.BetaApi;
 import com.google.api.core.InternalApi;
-import com.google.common.base.Throwables;
-import com.google.common.collect.Queues;
 import java.util.Iterator;
-import java.util.NoSuchElementException;
-import java.util.concurrent.BlockingQueue;
 import javax.annotation.Nonnull;
 
 /**
@@ -69,8 +65,6 @@ import javax.annotation.Nonnull;
  */
 @BetaApi("The surface for streaming is not stable yet and may change in the future.")
 public final class ServerStream<V> implements Iterable<V> {
-  private static final Object EOF_MARKER = new Object();
-
   private final QueuingResponseObserver<V> observer = new QueuingResponseObserver<>();
   private final ServerStreamIterator<V> iterator = new ServerStreamIterator<>(observer);
   private boolean consumed;
@@ -112,176 +106,5 @@ public final class ServerStream<V> implements Iterable<V> {
    */
   public void cancel() {
     observer.cancel();
-  }
-
-  /**
-   * Internal implementation of a blocking Iterator, which will coordinate with the
-   * QueuingResponseObserver fetch new items from upstream. The Iterator expects the observer to
-   * request the first item, afterwards, new items will be requested when the current ones are
-   * consumed by next().
-   *
-   * @param <V> The type of items to be Iterated over.
-   */
-  private static final class ServerStreamIterator<V> implements Iterator<V> {
-    private final QueuingResponseObserver<V> observer;
-    private Object last;
-
-    ServerStreamIterator(QueuingResponseObserver<V> observer) {
-      this.observer = observer;
-    }
-
-    /**
-     * Checks if the next call to {@code hasNext()} (and {@code next()}) will block.
-     *
-     * @return true if the next call is guaranteed to be nonblocking.
-     */
-    boolean isReady() {
-      return last != null || observer.isReady();
-    }
-
-    /**
-     * Consumes the next response and asynchronously request the next response from the observer.
-     *
-     * @return The next response.
-     * @throws NoSuchElementException If the stream has been consumed or cancelled.
-     */
-    @Override
-    public V next() {
-      if (!hasNext()) {
-        throw new NoSuchElementException();
-      }
-      try {
-        observer.request();
-        @SuppressWarnings("unchecked")
-        V tmp = (V) last;
-        return tmp;
-      } finally {
-        last = null;
-      }
-    }
-
-    /**
-     * Checks if the stream has been fully consumed or cancelled. This method will block until the
-     * observer enqueues another event (response or completion event). If the observer encountered
-     * an error, this method will propagte the error and put itself into an error where it will
-     * always return the same error.
-     *
-     * @return true If iterator has more responses.
-     */
-    @Override
-    public boolean hasNext() {
-      if (last == null) {
-        try {
-          last = observer.getNext();
-        } catch (InterruptedException e) {
-          Thread.currentThread().interrupt();
-          throw new RuntimeException(e);
-        }
-      }
-      if (last instanceof Throwable) {
-        Throwable throwable = (Throwable) this.last;
-
-        Throwables.throwIfUnchecked(throwable);
-        throw new RuntimeException(throwable);
-      }
-      return last != EOF_MARKER;
-    }
-
-    /** Unsupported. */
-    @Override
-    public void remove() {
-      throw new UnsupportedOperationException();
-    }
-  }
-
-  /**
-   * A back pressure aware bridge from a {@link ResponseObserver} to a {@link BlockingQueue}. The
-   * queue size is fixed to 1 item & a close signal. The observer will manage it's own flow control
-   * keeping the queue in one of 3 states:
-   *
-   * <ul>
-   *   <li>empty: a item has been requested and we are awaiting the next item
-   *   <li>1 item: an in progress stream with 1 item buffered
-   *   <li>1 control signal: either a Throwable or an EOF_MARKER means that the stream is closed
-   *   <li>1 item & 1 control signal: this is the last item of the stream
-   * </ul>
-   *
-   * The observer can also be abruptly cancelled, which cancels the underlying call and always
-   * returns an EOF_MARKER.
-   *
-   * @param <V> The item type.
-   */
-  private static final class QueuingResponseObserver<V> implements ResponseObserver<V> {
-    private final BlockingQueue<Object> buffer = Queues.newArrayBlockingQueue(2);
-    private StreamController controller;
-    private boolean isCancelled;
-
-    void request() {
-      controller.request(1);
-    }
-
-    Object getNext() throws InterruptedException {
-      if (isCancelled) {
-        return EOF_MARKER;
-      }
-      return buffer.take();
-    }
-
-    boolean isReady() {
-      return isCancelled || !buffer.isEmpty();
-    }
-
-    /**
-     * Cancels the underlying RPC and causes getNext to always return EOF_MARKER. This can only be
-     * called after starting the underlying call.
-     */
-    void cancel() {
-      isCancelled = true;
-      controller.cancel();
-    }
-
-    /**
-     * Before starting the RPC, disable automatic flow control and retain a reference to the
-     * controller.
-     *
-     * @param controller The controller for the stream.
-     */
-    @Override
-    public void onStart(StreamController controller) {
-      this.controller = controller;
-      controller.disableAutoInboundFlowControl();
-      controller.request(1);
-    }
-
-    /**
-     * Buffer the response. There should be at most 1 response in the buffer
-     *
-     * @param response The received response.
-     */
-    @Override
-    public void onResponse(V response) {
-      buffer.add(response);
-    }
-
-    /**
-     * Enqueue the error to be thrown later on. The error might occur w/o a request so the queue
-     * might grow to 2 elements, in that case the previous response will be consumed first.
-     *
-     * @param t The error occurred on the stream
-     */
-    @Override
-    public void onError(Throwable t) {
-      buffer.add(t);
-    }
-
-    /**
-     * Enqueue a marker to notify the consumer that the stream is finished. In most situations this
-     * will cause the queue to grow to 2 elements: the requested response and an unsolicited
-     * completion marker.
-     */
-    @Override
-    public void onComplete() {
-      buffer.add(EOF_MARKER);
-    }
   }
 }
